restruct into *_helper.go and *_unexported.go
