+++ pkg/kubernetes/pod/pod.go
@@ -1,7 +1,6 @@
 package pod
 
 import (
-	"bufio"
 	"context"
 	"fmt"
 	"strings"
@@ -16,22 +15,6 @@ import (
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/kubernetes"
 )
-
-// TODO: move to a helper.go file with all other functions that return anything but an error?
-func ListPodsWithLabelSelector(kubeClientset kubernetes.Interface, namespace, selector string) (*corev1.PodList, error) {
-	if err := common.ValidateClientset(kubeClientset); err != nil {
-		return nil, err
-	}
-
-	pods, err := util.RetryOnError(&util.DefaultRetry, util.IsRetriable, func() (interface{}, error) {
-		return kubeClientset.CoreV1().Pods(namespace).List(context.Background(), metav1.ListOptions{LabelSelector: selector})
-	})
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to list pods")
-	}
-
-	return pods.(*corev1.PodList), nil
-}
 
 func Pods(kubeClientset kubernetes.Interface, namespace string) error {
 	return PodsWithSelector(kubeClientset, namespace, "")
@@ -128,49 +111,6 @@ func SomeOrAllPodsInNamespaceWithSelectorHaveStringInLogsSinceTime(kubeClientset
 		}
 		return nil
 	})
-}
-
-func GetExpBackoff(steps int) wait.Backoff {
-	return wait.Backoff{
-		Duration: 2 * time.Second,
-		Factor:   2.0,
-		Jitter:   0.5,
-		Steps:    steps,
-		Cap:      10 * time.Minute,
-	}
-}
-
-func countStringInPodLogs(kubeClientset kubernetes.Interface, pod corev1.Pod, since time.Time, stringsToFind ...string) (int, error) {
-	foundCount := 0
-	if err := common.ValidateClientset(kubeClientset); err != nil {
-		return foundCount, err
-	}
-	var sinceTime metav1.Time = metav1.NewTime(since)
-	for _, container := range pod.Spec.Containers {
-		podLogOpts := corev1.PodLogOptions{
-			SinceTime: &sinceTime,
-			Container: container.Name,
-		}
-
-		req := kubeClientset.CoreV1().Pods(pod.Namespace).GetLogs(pod.Name, &podLogOpts)
-		podLogs, err := req.Stream(context.Background())
-		if err != nil {
-			return 0, errors.Errorf("Error in opening stream for pod '%s', container '%s' : '%s'", pod.Name, container.Name, string(err.Error()))
-		}
-
-		scanner := bufio.NewScanner(podLogs)
-		for scanner.Scan() {
-			line := scanner.Text()
-			for _, stringToFind := range stringsToFind {
-				if strings.Contains(line, stringToFind) {
-					foundCount += 1
-					log.Infof("Found string '%s' in line '%s' in container '%s' of pod '%s'", stringToFind, line, container.Name, pod.Name)
-				}
-			}
-		}
-		podLogs.Close()
-	}
-	return foundCount, nil
 }
 
 func SomePodsInNamespaceWithSelectorDontHaveStringInLogsSinceTime(kubeClientset kubernetes.Interface, namespace, selector, searchkeyword string, since time.Time) error {
